poc(modal): make the video timeline uni-modal, drop the multimodal shell
Video datasets are uni-modal; the shell was the wrong shape for them.
Revert every multimodal change (root export, components barrel, exports
entry) and rebuild the POC directly on @fiftyone/playback: player on top,
TimelineWithTracks beneath, no tiling, no header, no sidebars.

playback and tiling carry no multimodal dependency — only
MultiModalPlayback is the multimodal-branded composition — so the video
surface composes the generic packages and owes multimodal nothing.

Tiling stays in scope for the *group* case only, where the same timeline
gains one tile per selected video slice on a shared clock. Not built here.
